requirements.txt
streamlit==1.28.1
pandas==2.1.3
google-generativeai==0.3.0
openpyxl==3.11.0
